refactor: Change label.infra packages
- 패키지 이름 수정
